+++ benchmark/Benchmark.hs
@@ -7,7 +7,6 @@ import Data.Vector.Storable as VS
 import Data.Vector.Unboxed as VU
 
 import Data.Vec3
-import qualified Data.Vec3.Unboxed.Contiguous as C
 
 
 oXYZ :: (Double, Double, Double)
@@ -52,11 +51,11 @@ testDotM v v2 = whnf (uncurry $ VG.zipWith dotM') (v, v2)
 
 -- Note that source arrays are not forced in test functions.
 
-tv :: VU.Vector UVec3
+tv :: VU.Vector TUVec3
 tv = VG.replicate n $ fromXYZ oXYZ
 
 
-cv :: VU.Vector C.UVec3
+cv :: VU.Vector UVec3
 cv = VG.replicate n $ fromXYZ oXYZ
 
 
@@ -64,11 +63,11 @@ sv :: VS.Vector SVec3
 sv = VG.replicate n $ fromXYZ oXYZ
 
 
-tv' :: VU.Vector UVec3
+tv' :: VU.Vector TUVec3
 tv' = VG.replicate n $ fromXYZ oXYZ'
 
 
-cv' :: VU.Vector C.UVec3
+cv' :: VU.Vector UVec3
 cv' = VG.replicate n $ fromXYZ oXYZ'
 
 
@@ -80,11 +79,11 @@ bsv :: VS.Vector SVec3
 bsv = VG.replicate bigN $ fromXYZ oXYZ
 
 
-bcv :: VU.Vector C.UVec3
+bcv :: VU.Vector UVec3
 bcv = VG.replicate bigN $ fromXYZ oXYZ
 
 
-btv :: VU.Vector UVec3
+btv :: VU.Vector TUVec3
 btv = VG.replicate bigN $ fromXYZ oXYZ
 
 
@@ -92,11 +91,11 @@ bsv' :: VS.Vector SVec3
 bsv' = VG.replicate bigN $ fromXYZ oXYZ'
 
 
-bcv' :: VU.Vector C.UVec3
+bcv' :: VU.Vector UVec3
 bcv' = VG.replicate bigN $ fromXYZ oXYZ'
 
 
-btv' :: VU.Vector UVec3
+btv' :: VU.Vector TUVec3
 btv' = VG.replicate bigN $ fromXYZ oXYZ'
 
 
@@ -104,32 +103,32 @@ main :: IO ()
 main = defaultMain
        [ bgroup "zipWith <+>"
                     [ bench "SVec/Storable"           $ testWhole sv sv'
-                    , bench "UVec/Unboxed"            $ testWhole tv tv'
+                    , bench "UVec/Unboxed/Tupled"     $ testWhole tv tv'
                     , bench "UVec/Unboxed/Contiguous" $ testWhole cv cv'
                     ]
        , bgroup "zipWith-by-x +"
                     [ bench "SVec/Storable"           $ testComp sv sv'
-                    , bench "UVec/Unboxed"            $ testComp tv tv'
+                    , bench "UVec/Unboxed/Tupled"     $ testComp tv tv'
                     , bench "UVec/Unboxed/Contiguous" $ testComp cv cv'
                     ]
        , bgroup "zipWith dotM"
                     [ bench "SVec/Storable"           $ testDotM sv sv'
-                    , bench "UVec/Unboxed"            $ testDotM tv tv'
+                    , bench "UVec/Unboxed/Tupled"     $ testDotM tv tv'
                     , bench "UVec/Unboxed/Contiguous" $ testDotM cv cv'
                     ]
        , bgroup "zipWith <+> 1M"
                     [ bench "SVec/Storable"           $ testWhole bsv bsv'
-                    , bench "UVec/Unboxed"            $ testWhole btv btv'
+                    , bench "UVec/Unboxed/Tupled"     $ testWhole btv btv'
                     , bench "UVec/Unboxed/Contiguous" $ testWhole bcv bcv'
                     ]
        , bgroup "zipWith-by-x + 1M"
                     [ bench "SVec/Storable"           $ testComp bsv bsv'
-                    , bench "UVec/Unboxed"            $ testComp btv btv'
+                    , bench "UVec/Unboxed/Tupled"     $ testComp btv btv'
                     , bench "UVec/Unboxed/Contiguous" $ testComp bcv bcv'
                     ]
        , bgroup "zipWith dotM 1M"
                     [ bench "SVec/Storable"           $ testDotM bsv bsv'
-                    , bench "UVec/Unboxed"            $ testDotM btv btv'
+                    , bench "UVec/Unboxed/Tupled"     $ testDotM btv btv'
                     , bench "UVec/Unboxed/Contiguous" $ testDotM bcv bcv'
                     ]
        ]